% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{create_plots}
\alias{create_plots}
\title{Create EDA plots and tables}
\usage{
create_plots(out_dir)
}
\arguments{
\item{out_dir}{A path to a directory.}
}
\value{
out_dir (invisible).
}
\description{
This function creates the figures and tables of the Exploratory Data
Analysis of the Amazonia deforestation warnings issued by DETER.
}
\examples{
\dontrun{
    create_plots(out_dir = "~/Documents/report")
}
}
