% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{filter_trajectory}
\alias{filter_trajectory}
\title{Filter trajectories}
\usage{
filter_trajectory(
  data_tb,
  filter_col,
  filter_val,
  said_col = "xy_id",
  invert = FALSE
)
}
\arguments{
\item{data_tb}{A tibble with trajectory data.}

\item{filter_col}{A character. Name of a column.}

\item{filter_val}{Value to filter.}

\item{said_col}{A character. Name of the column with subarea ids.}

\item{invert}{A logical. Should the filter results be inverted?}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A tibble.
}\if{html}{\out{</div>}}
}
\description{
This function filters whole trajectories based on attributes of a single
trajectory step.
}
