% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{flat_trajectory}
\alias{flat_trajectory}
\title{Flat trajectories}
\usage{
flat_trajectory(
  data_tb,
  keep = "first",
  said_col = "xy_id",
  year_col = "year",
  date_col = "VIEW_DATE",
  other_col = NA
)
}
\arguments{
\item{data_tb}{A tibble with trajectory data.}

\item{keep}{A character. Obervation to keep. Either "first" or "last".}

\item{said_col}{A character. Name of the column with subarea ids.}

\item{year_col}{A character. Name of the column with the observations'
year.}

\item{date_col}{A character. Name of the column with the observations'
date.}

\item{other_col}{A character. Name of an additional grouping column.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A tibble.
}\if{html}{\out{</div>}}
}
\description{
This function ensures there is only one event in each year.
}
