% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter.R
\name{get_deter_classes}
\alias{get_deter_classes}
\title{Get the DETER classes}
\usage{
get_deter_classes()
}
\value{
A tibble.
}
\description{
This function returns a tibble with the classes in the DETER data.
}
