% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subarea.R
\name{get_flat_subarea}
\alias{get_flat_subarea}
\title{Return an flat subarea SF object}
\usage{
get_flat_subarea(subarea_sf)
}
\arguments{
\item{subarea_sf}{An sf object.}
}
\value{
An sf object.
}
\description{
This function removes overlaped subareas.
}
