% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_area_by_class_state}
\alias{get_plot_area_by_class_state}
\title{Build a histogram using the classes of DETER by Brazilian state.}
\usage{
get_plot_area_by_class_state(subarea_dt)
}
\arguments{
\item{subarea_dt}{A data.table object hosted in this package.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a histogram figure using the classes of the
deforestation warnings by subarea and state in the brazilian amazon. The
figure shows the area by DETER class, and year (PRODES).
}
