% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_area_by_state_year_type}
\alias{get_plot_area_by_state_year_type}
\title{Build a treemap using of DETER's deforestation warnings.}
\usage{
get_plot_area_by_state_year_type(subarea_dt)
}
\arguments{
\item{subarea_dt}{A data.table object hosted in this package.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a TreeMap figure using deforestation warnings in the
brazilian amazon. The figure uses DETER warnings area aggregated by state,
PRODES year, and warning's type and area.
}
