% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_area_by_warnings_state}
\alias{get_plot_area_by_warnings_state}
\title{Build histograms by state using DETER deforestation warnings.}
\usage{
get_plot_area_by_warnings_state(
  subarea_tb,
  area_breaks,
  area_col = "subarea_ha",
  said_col = "xy_id",
  state_col = "UF"
)
}
\arguments{
\item{subarea_tb}{A tibble hosted in this package.}

\item{area_breaks}{A named numeric vector with the breaks of subareas by
size.}

\item{area_col}{A character. Name of the column with area (ha).}

\item{said_col}{A character. Nma of the subarea ID column.}

\item{state_col}{A character. Name of the column with state.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a histogram figure using deforestation warnings
subareas for each state in the brazilian amazon. The figure shows the number
of warnnings by subarea, state,  and subarea bins.
}
