% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_area_ndays}
\alias{get_plot_area_ndays}
\alias{get_plot_density_area_ndays}
\title{Build a dot cloud figure using DETER deforestation warnings.}
\usage{
get_plot_density_area_ndays(
  subarea_tb,
  area_col = "subarea_ha",
  date_col = "VIEW_DATE",
  label_col = "CLASSNAME",
  said_col = "xy_id",
  state_col = "UF",
  year_col = "year"
)
}
\arguments{
\item{subarea_tb}{A tibble hosted in this package.}

\item{area_col}{A character. Name of the column with area (ha).}

\item{date_col}{A character. Name of the view date column.}

\item{label_col}{A character. Name of the column with labels.}

\item{said_col}{A character. Nma of the subarea ID column.}

\item{state_col}{A character. Name of the column with state.}

\item{year_col}{A character. Name of the column with year (PRODES).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a dot cloud and density figure using deforestation
warnings subareas in the brazilian amazon. The figure uses DETER warnings
organized by number of days between warnings, area, warning class (of the
first warning), state, and PRODES year. Only subareas greater then 3 ha are
included.
}
