% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_days_first_to_last}
\alias{get_plot_days_first_to_last}
\title{Build boxplots of the days from first to last DETER warning.}
\usage{
get_plot_days_first_to_last(subarea_dt, area_breaks)
}
\arguments{
\item{subarea_dt}{A data.table object hosted in this package.}

\item{area_breaks}{A named numeric vector with the breaks of subareas by
size.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates figure of boxplots by state and number of warnings of
number of days between the first and last DETER deforestation warning
associated to each subarea.
}
