% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_fire_by_month}
\alias{get_plot_fire_by_month}
\title{Build a histogram plot of fire spot data}
\usage{
get_plot_fire_by_month(
  data_tb,
  date_col = "datahora",
  year_col = "year",
  month_col = "month"
)
}
\arguments{
\item{data_tb}{A tibble.}

\item{date_col}{A character. Name of the column with date time of fire
spot.}

\item{year_col}{A character. Name of the column with year.}

\item{month_col}{A character. Name of the column with month.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a histogram and line plot of the given fire spot data.
}
