% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_sankey_year}
\alias{get_plot_sankey_year}
\title{Build a Sankey diagram of given subareas using their year}
\usage{
get_plot_sankey_year(
  data_tb,
  area_col = "subarea_ha",
  label_col = "CLASSNAME",
  said_col = "xy_id",
  year_col = "year"
)
}
\arguments{
\item{data_tb}{A tibble.}

\item{area_col}{A character. Name of the column with area (ha).}

\item{label_col}{A character. Name of the column with labels.}

\item{said_col}{A character. Nma of the subarea ID column.}

\item{year_col}{A character. Name of the column with year (PRODES).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a Sankey diagram between DETER subareas for the given
number of warnings using their PRODES year.
}
