% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{get_prodes_codes}
\alias{get_prodes_codes}
\title{Get the PRODES codes.}
\usage{
get_prodes_codes()
}
\value{
A tibble.
}
\description{
This function returns the numeric codes using for the PRODES classes. This
function keeps track of the classes used in PRODES products (vector and
raster) along time.
}
