% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{get_prodes_names}
\alias{get_prodes_names}
\title{Prepare a vector with PRODES classes.}
\usage{
get_prodes_names()
}
\value{
An character.
}
\description{
This function returns a named vector with PRODES common names and their
corresponding classes from the raster and vector PRODES data. This vector
is useful for recoding.
}
