% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{process_prodes_codes}
\alias{process_prodes_codes}
\title{Extract the PRODES codes from text file}
\usage{
process_prodes_codes(file_path)
}
\arguments{
\item{file_path}{Path to the text file downloaded along with the PRODES
raster.}
}
\value{
A tibble with two columns: code, and class.
}
\description{
This function extracts the codes in the text file associated to the PRODES
raster of a given year.
}
