% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{report_suspicious_trajectories}
\alias{report_suspicious_trajectories}
\title{Find potential inconsistent trajectories}
\usage{
report_suspicious_trajectories(subarea_sf, subarea_dt, out_dir)
}
\arguments{
\item{subarea_sf}{An sf object with data about trajectories. See the data
in this package.}

\item{subarea_dt}{A data.table object with data about subtrajectories. See
the data in this package.}

\item{out_dir}{A path to a directory.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A character with paths to files storing trajectories which
           start with the class deforestation or end with the class
           forest.
}\if{html}{\out{</div>}}
}
\description{
This function reports trajectories with potential inconsistencies.
of steps.
}
