% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subarea.R
\name{test_centroids}
\alias{test_centroids}
\title{Compare subareas' id to their centroids}
\usage{
test_centroids(data_sf, sample_n = 50, tol = 1e-05)
}
\arguments{
\item{data_sf}{An sf object.}

\item{sample_n}{An integer.}

\item{tol}{A numeric. The tolerance used for the comparison.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A logical.
}\if{html}{\out{</div>}}
}
\description{
This function compares the subarea id, which is made of the centroids of
each subarea, to the actual centroids computed using the sf package.
}
