% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{trim_trajectories}
\alias{trim_trajectories}
\title{End the trajectories at certain class.}
\usage{
trim_trajectories(
  data_tb,
  end_class,
  said_col = "xy_id",
  label_col = "CLASSNAME",
  date_col = "VIEW_DATE"
)
}
\arguments{
\item{data_tb}{A tibble with trajectory data.}

\item{end_class}{A character. Classes on which trajectories should end.}

\item{said_col}{A character. Name of the column with}

\item{label_col}{A character. Name of the column with labels.}

\item{date_col}{A character. Name of the view date column.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     A tibble.
}\if{html}{\out{</div>}}
}
\description{
This function computes removes events from trajectories after certain class
is found.
}
