% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMOVE_eda.R, R/exploratory_data_analysis.R
\name{create_plots}
\alias{create_plots}
\title{Create EDA plots and tables}
\usage{
create_plots(out_dir, subarea_tb, fire_sf)

create_plots(out_dir, subarea_tb, fire_sf)
}
\arguments{
\item{out_dir}{A path to a directory.}

\item{subarea_tb}{A tibble with data about DETER subareas.}

\item{fire_sf}{A sf object (point). Fire spots.}
}
\value{
out_dir (invisible).

out_dir (invisible).
}
\description{
This function creates the figures and tables of the Exploratory Data
Analysis of the Amazonia deforestation warnings issued by DETER.

This function creates the figures and tables of the Exploratory Data
Analysis of the Amazonia deforestation warnings issued by DETER.
}
\examples{
\dontrun{
    create_plots(out_dir = "~/Documents/report")
}
\dontrun{
    create_plots(out_dir = "~/Documents/report")
}
}
