% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_area_by_class_state}
\alias{get_plot_area_by_class_state}
\title{Build a histogram using the classes of DETER by Brazilian state.}
\usage{
get_plot_area_by_class_state(
  subarea_tb,
  area_col = "subarea_ha",
  label_col = "CLASSNAME",
  state_col = "UF",
  year_col = "year"
)
}
\arguments{
\item{subarea_tb}{A tibble hosted in this package.}

\item{area_col}{A character. Name of the column with area (ha).}

\item{label_col}{A character. Name of the column with labels.}

\item{state_col}{A character. Name of the column with state.}

\item{year_col}{A character. Name of the column with year (PRODES).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a histogram figure using the classes of the
deforestation warnings by subarea and state in the brazilian amazon. The
figure shows the area by DETER class, and year (PRODES).
}
