% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_area_by_state_year_type}
\alias{get_plot_area_by_state_year_type}
\title{Build a treemap using of DETER's deforestation warnings.}
\usage{
get_plot_area_by_state_year_type(
  subarea_tb,
  class_levels,
  area_col = "subarea_ha",
  label_col = "CLASSNAME",
  state_col = "UF",
  year_col = "year"
)
}
\arguments{
\item{subarea_tb}{A tibble hosted in this package.}

\item{class_levels}{A character. Names used to recode DETER's CLASSNAME.}

\item{area_col}{A character. Name of the column with area (ha).}

\item{label_col}{A character. Name of the column with labels.}

\item{state_col}{A character. Name of the column with state.}

\item{year_col}{A character. Name of the column with year (PRODES).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a TreeMap figure using deforestation warnings in the
brazilian amazon. The figure uses DETER warnings area aggregated by state,
PRODES year, and warning's type and area.
}
