% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_area_by_warnings_state}
\alias{get_plot_area_by_warnings_state}
\title{Build histograms by state using DETER deforestation warnings.}
\usage{
get_plot_area_by_warnings_state(subarea_dt, area_breaks)
}
\arguments{
\item{subarea_dt}{A data.table object hosted in this package.}

\item{area_breaks}{A named numeric vector with the breaks of subareas by
size.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a histogram figure using deforestation warnings
subareas for each state in the brazilian amazon. The figure shows the number
of warnnings by subarea, state,  and subarea bins.
}
