% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_area_ndays}
\alias{get_plot_area_ndays}
\alias{get_plot_density_area_ndays}
\title{Build a dot cloud figure using DETER deforestation warnings.}
\usage{
get_plot_density_area_ndays(subarea_dt)
}
\arguments{
\item{subarea_dt}{A data.table object hosted in this package.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a dot cloud and density figure using deforestation
warnings subareas in the brazilian amazon. The figure uses DETER warnings
organized by number of days between warnings, area, warning class (of the
first warning), state, and PRODES year. Only subareas greater then 3 ha are
included.
}
