% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_sankey}
\alias{get_plot_sankey}
\title{Build a Sankey diagram of DETER warnings.}
\usage{
get_plot_sankey(data_tb, n_warnings)
}
\arguments{
\item{data_tb}{A data.table object specifically prepared for this function.}

\item{n_warnings}{A single integer. Number of warnings to use.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       A ggplot2 object.
}\if{html}{\out{</div>}}
}
\description{
This function creates a Sankey diagram between DETER subareas for the given
number of warnings.
}
