% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{get_trajectory_stats}
\alias{get_trajectory_stats}
\title{Compute trajectory statistics}
\usage{
get_trajectory_stats(data_tb)
}
\arguments{
\item{data_tb}{A tibble with trajectory data.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A tibble.
}\if{html}{\out{</div>}}
}
\description{
This function computes statistics about trajectories with the same number
of steps.
}
