% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R, R/util.R
\name{the_mode}
\alias{the_mode}
\alias{get_prodes_codes}
\title{Get the PRODES codes.}
\usage{
get_prodes_codes()

the_mode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A tibble.

A vector.
}
\description{
This function returns the numeric codes using for the PRODES classes. This
function keeps track of the classes used in PRODES products along the years
(raster and vector).

This function computes the mode of the given vector, ignoring NA elements.
}
\examples{
the_mode(sample(1:10, size = 100, replace = TRUE))
}
