% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{the_mode}
\alias{the_mode}
\title{Compute the mode of a vector.}
\usage{
the_mode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector.
}
\description{
This function computes the mode of the given vector, ignoring NA elements.
}
\examples{
the_mode(sample(1:10, size = 100, replace = TRUE))
}
