% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{trim_trajectories}
\alias{trim_trajectories}
\title{End the trajectories at certain class.}
\usage{
trim_trajectories(data_tb, end_class)
}
\arguments{
\item{data_tb}{A tibble with trajectory data.}

\item{end_class}{A character. Classes on which trajectories should end.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     A tibble.
}\if{html}{\out{</div>}}
}
\description{
This function computes removes events from trajectories after certain class
is found.
}
